(dev) remove inundation calculations from the repository
The inundation calculations were initially featured in the repository
in order to develop a full fully featured tool for reservoir and
catchment delineation followed by greenhouse gas emission estimation.
All calculations were supposed to be performed offline on a local
machine. Currently dam-emission is developed in separation to
reservoir and catchment delineation

The inundation code originally featured in this repository has been
moved to a new repository:
https://github.com/tomjanus/dam-inundation