(doc, refactor, fix): Update MonthlyTemperarure.
- Documents code.
- Fixes discrepancy with g-Res calculation method.
- Changes argument to eff_temp() method from coefficient to gas type.

Effective temperature was previously calculated with a single equation
for all temperatures. The code is now changed to follow g-Res equation
in which temperatures below 4 deg C are treated as 4 deg C. I.e.
there is a min. cap of 4 deg C for all temperature values in the
temperature profile.