fix: Remove capital letters from climate categories in .yaml config files.
Fixes the issue of producing zero pre-impoundment emissions for reservoirs
in which the pre-impoundment emissions would otherwise not be zero. The
error was due to the fact that the code tried to obtain values for keys
that did not exist in the config files, e.g. key 'boreal' while the
key in the config file was 'Boreal'. This, coupled with the `.get` methods
for accessing dictionary values with a default value of zero resulted in
zero coefficients produced for non-matched keys.