feat: Add compute P exports using reduced catchment area
Adds an option to derive landuse composition and areas for
reservoir catchment post-inundiation, i.e. when
catchment := catchment - reservoir.

Provides an a factory class which takes the (original) catchment
object and the reservoir object as arguments and produces the
reduced catchment object based on which (reduced) P expoerts
can be calculated.