feat: Expose internal vars as outputs. Fixes #33 and #25
Adds a decorator that, when applied to a method, adds its return
value to a dictionary of internal variables (a global variable
declared in `globals.py`.

Uses a `yaml` file to define the names, units, descriptions and
include flags of the internal variables exposed from within the
source code.

Includes internal variables in the data presentation logic of all
three writers: json, latex/pdf and excel.

Adds a wrapper for latex->pdf compilation code in `pylatex` that
allows multiple compilations in a row. Two compilations in a row
are required for the calculation of the total number of pages that
are displayed in the header of the pdf report. Hence, it fixes
issue #25.