feat: Correct units in N2O emission equations, revert #73a5f2c
In #73a5f2c we multiplied equations by factor of 1e3 to convert
g into kg. We now realised that the units were already in kgs and
therefore, additional conversion was not needed. Ultimately, with
the factor of 1e3 the values were 1,000 too high. This commit
reverts this previous change.