chore: Change references to heet into geocaret
Reflects change of names of the 'upstream' package which produces
inputs to re-emission by delineating reservoir and catchment and
querying the clipped reservoir and catchment polygons with spatial
datasets including geomorphological and climatic information about
the reservoir and the catchment.

Previously the package was called HEET. It's been recently renamed
to GeoCARET: "Geospatial CAtchment and REservoir analysis Tool.