feat(vis): enable report generation in HTML format
Adds HTMLWriter class which creates HTML reports that visually resemble and
share some of the code with LatexWriter that is used for generating PDF
reports. HTMLWriter.write() is automatically called in CLI if the output
file format has a .html extension.

HTML writer was added to enable report generation capability without requiring
a bulky LaTeX compiler that has so far been required for generating PDF reports.