dev (salib): Add sensitivity analysis capability with SALib package.
Enables the use of SALib for performing global sensitivity analysis
to model parameters and inputs.

Model parameters are exposed using the dynamic configuration capability
that allows reading and modifying parameters 'on-the-fly'. Model inputs
are modifying by accessing the Input objects for each specific scenario
(reservoir).

The module currently supports analysis of sensitivity (e.g. for uncertainty
quantification) using the Sobol method, but can be extended to other methdods
supported in SALib. It provides a wrapper for re-emission that enables running
the sensitivity for individual reservoir (single-scenario) and batches of multiple
reservoirs (multiple-scenario). It also provides several plot types for visualizing
the sensitivity in single-scenario and multiple-scenario settings.